refactor(agents): split chat-sdk into ChatInstanceRegistry + InboundDispatcher + gateway; delete chat-sdk.service
