+++ apps/api/src/app/agents/conversation-runtime/egress/outbound.gateway.ts
@@ -1,10 +1,14 @@
-import { Injectable } from '@nestjs/common';
+import { BadGatewayException, BadRequestException, Injectable } from '@nestjs/common';
 import { ConversationChannel } from '@novu/dal';
 import type { SentMessageInfo } from '@novu/framework';
-import type { Thread } from 'chat';
+import type { AdapterPostableMessage, EmojiValue, PlanModel, Thread } from 'chat';
+import { AgentConfigResolver } from '../../services/agent-config-resolver.service';
 import { AgentConversationService } from '../../services/agent-conversation.service';
-import { ChatSdkService } from '../../services/chat-sdk.service';
 import type { ReplyContentDto } from '../../shared/dtos/agent-reply-payload.dto';
+import { esmImport } from '../../shared/util/esm-import';
+import { ChatInstanceRegistry } from '../ingress/chat-instance.registry';
+import type { ChatSdkFile, ChatSdkReplyContent } from './file-materializer.service';
+import { FileMaterializer } from './file-materializer.service';
 
 export interface ConversationTarget {
   agentId: string;
@@ -24,11 +28,45 @@ export interface OutboundPersistContext {
 
 export type OutboundMessage = ReplyContentDto;
 
+function getErrorResponseBody(err: unknown): unknown {
+  if (!err || typeof err !== 'object') {
+    return undefined;
+  }
+
+  return (err as { response?: { body?: unknown } }).response?.body;
+}
+
+function getDeliveryErrorDetail(body: unknown): string | undefined {
+  if (!body || typeof body !== 'object') {
+    return undefined;
+  }
+
+  const responseBody = body as { errors?: Array<{ message?: unknown }>; message?: unknown };
+  const firstErrorMessage = responseBody.errors?.[0]?.message;
+  if (typeof firstErrorMessage === 'string') {
+    return firstErrorMessage;
+  }
+
+  return typeof responseBody.message === 'string' ? responseBody.message : undefined;
+}
+
+function toDeliveryError(err: unknown): never {
+  const base = err instanceof Error ? err.message : String(err);
+  const detail = getDeliveryErrorDetail(getErrorResponseBody(err));
+
+  throw new BadGatewayException({
+    error: 'delivery_failed',
+    message: detail ? `${base}: ${detail}` : base,
+  });
+}
+
 @Injectable()
 export class OutboundGateway {
   constructor(
-    private readonly chat: ChatSdkService,
-    private readonly conversation: AgentConversationService
+    private readonly registry: ChatInstanceRegistry,
+    private readonly conversation: AgentConversationService,
+    private readonly agentConfigResolver: AgentConfigResolver,
+    private readonly fileMaterializer: FileMaterializer
   ) {}
 
   async deliver(
@@ -36,7 +74,7 @@ export class OutboundGateway {
     msg: OutboundMessage,
     persist: OutboundPersistContext
   ): Promise<SentMessageInfo> {
-    const sent = await this.chat.postToConversation(
+    const sent = await this.postToConversation(
       target.agentId,
       target.integrationIdentifier,
       target.platform,
@@ -54,7 +92,7 @@ export class OutboundGateway {
     msg: OutboundMessage,
     persist: OutboundPersistContext
   ): Promise<SentMessageInfo> {
-    const sent = await this.chat.editInConversation(
+    const sent = await this.editInConversation(
       target.agentId,
       target.integrationIdentifier,
       target.platform,
@@ -96,6 +134,208 @@ export class OutboundGateway {
 
       throw err;
     }
+  }
+
+  async postToConversation(
+    agentId: string,
+    integrationIdentifier: string,
+    platform: string,
+    platformThreadId: string,
+    content: ReplyContentDto
+  ): Promise<SentMessageInfo> {
+    const config = await this.agentConfigResolver.resolve(agentId, integrationIdentifier);
+    const instanceKey = `${agentId}:${integrationIdentifier}`;
+    const chat = await this.registry.getOrCreate(instanceKey, agentId, config.platform, config);
+
+    const thread = chat.thread(platformThreadId);
+    const deliveryContent = await this.fileMaterializer.prepareContentForDelivery(content, platform, agentId);
+
+    const postArg = this.buildAdapterPostableMessage(deliveryContent);
+
+    const sent = await thread.post(postArg).catch(toDeliveryError);
+
+    return { messageId: sent.id, platformThreadId: sent.threadId };
+  }
+
+  async startTypingInConversation(
+    agentId: string,
+    integrationIdentifier: string,
+    platformThreadId: string,
+    status = 'Thinking...'
+  ): Promise<void> {
+    const config = await this.agentConfigResolver.resolve(agentId, integrationIdentifier);
+    const instanceKey = `${agentId}:${integrationIdentifier}`;
+    const chat = await this.registry.getOrCreate(instanceKey, agentId, config.platform, config);
+    const thread = chat.thread(platformThreadId);
+
+    if (typeof thread.startTyping !== 'function') {
+      return;
+    }
+
+    await thread.startTyping(status).catch(toDeliveryError);
+  }
+
+  async sendDirectMessage(
+    agentId: string,
+    integrationIdentifier: string,
+    platformUserId: string,
+    content: ReplyContentDto
+  ): Promise<SentMessageInfo> {
+    const config = await this.agentConfigResolver.resolve(agentId, integrationIdentifier);
+    const instanceKey = `${agentId}:${integrationIdentifier}`;
+    const chat = await this.registry.getOrCreate(instanceKey, agentId, config.platform, config);
+
+    const dmThread = await chat.openDM(platformUserId);
+    const deliveryContent = await this.fileMaterializer.prepareContentForDelivery(content, config.platform, agentId);
+
+    const postArg = this.buildAdapterPostableMessage(deliveryContent);
+
+    const sent = await dmThread.post(postArg).catch(toDeliveryError);
+
+    const platformThreadId = sent.threadId.endsWith(':') ? `${sent.threadId}${sent.id}` : sent.threadId;
+
+    return { messageId: sent.id, platformThreadId };
+  }
+
+  async editInConversation(
+    agentId: string,
+    integrationIdentifier: string,
+    platform: string,
+    platformThreadId: string,
+    platformMessageId: string,
+    content: ReplyContentDto
+  ): Promise<SentMessageInfo> {
+    const config = await this.agentConfigResolver.resolve(agentId, integrationIdentifier);
+    const instanceKey = `${agentId}:${integrationIdentifier}`;
+    const chat = await this.registry.getOrCreate(instanceKey, agentId, config.platform, config);
+
+    const adapter = chat.getAdapter(platform);
+    if (typeof adapter.editMessage !== 'function') {
+      throw new BadRequestException(`Platform ${platform} does not support editing messages`);
+    }
+
+    const deliveryContent = await this.fileMaterializer.prepareContentForDelivery(content, platform, agentId);
+
+    const editPayload = this.buildAdapterPostableMessage(deliveryContent);
+
+    let editPromise: Promise<{ id: string; threadId: string }>;
+    if (deliveryContent.card) {
+      editPromise = adapter.editMessage(
+        platformThreadId,
+        platformMessageId,
+        deliveryContent.card as unknown as AdapterPostableMessage
+      );
+    } else {
+      editPromise = adapter.editMessage(platformThreadId, platformMessageId, editPayload);
+    }
+
+    const edited = await editPromise.catch(toDeliveryError);
+
+    return { messageId: edited.id, platformThreadId: edited.threadId };
+  }
+
+  async postPlanObject(
+    agentId: string,
+    integrationIdentifier: string,
+    platform: string,
+    platformThreadId: string,
+    model: PlanModel
+  ): Promise<SentMessageInfo | null> {
+    const config = await this.agentConfigResolver.resolve(agentId, integrationIdentifier);
+    const instanceKey = `${agentId}:${integrationIdentifier}`;
+    const chat = await this.registry.getOrCreate(instanceKey, agentId, config.platform, config);
+
+    const adapter = chat.getAdapter(platform);
+    if (typeof adapter.postObject !== 'function') {
+      return null;
+    }
+
+    const sent = await adapter.postObject(platformThreadId, 'plan', model).catch(toDeliveryError);
+
+    return { messageId: sent.id, platformThreadId: sent.threadId };
+  }
+
+  async editPlanObject(
+    agentId: string,
+    integrationIdentifier: string,
+    platform: string,
+    platformThreadId: string,
+    platformMessageId: string,
+    model: PlanModel
+  ): Promise<void> {
+    const config = await this.agentConfigResolver.resolve(agentId, integrationIdentifier);
+    const instanceKey = `${agentId}:${integrationIdentifier}`;
+    const chat = await this.registry.getOrCreate(instanceKey, agentId, config.platform, config);
+
+    const adapter = chat.getAdapter(platform);
+    if (typeof adapter.editObject !== 'function') {
+      return;
+    }
+
+    await adapter.editObject(platformThreadId, platformMessageId, 'plan', model).catch(toDeliveryError);
+  }
+
+  async reactToMessage(
+    agentId: string,
+    integrationIdentifier: string,
+    platform: string,
+    platformThreadId: string,
+    platformMessageId: string,
+    emojiName: string
+  ): Promise<void> {
+    const config = await this.agentConfigResolver.resolve(agentId, integrationIdentifier);
+    const instanceKey = `${agentId}:${integrationIdentifier}`;
+    const chat = await this.registry.getOrCreate(instanceKey, agentId, config.platform, config);
+
+    const adapter = chat.getAdapter(platform);
+    const resolved = await this.resolveEmoji(emojiName);
+    await adapter.addReaction(platformThreadId, platformMessageId, resolved);
+  }
+
+  async removeReaction(
+    agentId: string,
+    integrationIdentifier: string,
+    platform: string,
+    platformThreadId: string,
+    platformMessageId: string,
+    emojiName: string
+  ): Promise<void> {
+    const config = await this.agentConfigResolver.resolve(agentId, integrationIdentifier);
+    const instanceKey = `${agentId}:${integrationIdentifier}`;
+    const chat = await this.registry.getOrCreate(instanceKey, agentId, config.platform, config);
+
+    const adapter = chat.getAdapter(platform);
+    const resolved = await this.resolveEmoji(emojiName);
+    await adapter.removeReaction(platformThreadId, platformMessageId, resolved);
+  }
+
+  private async resolveEmoji(name: string): Promise<EmojiValue> {
+    const { getEmoji } = await esmImport('chat');
+    const resolved = getEmoji(name);
+    if (!resolved) {
+      throw new Error(`Unknown emoji name: "${name}". Use GET /agents/emoji to list supported options.`);
+    }
+
+    return resolved;
+  }
+
+  private buildAdapterPostableMessage(deliveryContent: ChatSdkReplyContent): AdapterPostableMessage {
+    if (deliveryContent.card) {
+      const payload: { card: unknown; files?: ChatSdkFile[] } = {
+        card: deliveryContent.card,
+      };
+
+      if (deliveryContent.files?.length) {
+        payload.files = deliveryContent.files;
+      }
+
+      return payload as unknown as AdapterPostableMessage;
+    }
+
+    return {
+      markdown: deliveryContent.markdown ?? '',
+      files: deliveryContent.files,
+    } as unknown as AdapterPostableMessage;
   }
 
   private async persistDelivered(
