+++ apps/api/src/app/agents/e2e/agent-reply.e2e.ts
@@ -2,8 +2,8 @@ import { ConversationActivitySenderTypeEnum, ConversationActivityTypeEnum, Conve
 import { testServer } from '@novu/testing';
 import { expect } from 'chai';
 import sinon from 'sinon';
+import { OutboundGateway } from '../conversation-runtime/egress/outbound.gateway';
 import { AgentExecutionParams, BridgeExecutorService } from '../services/bridge-executor.service';
-import { ChatSdkService } from '../services/chat-sdk.service';
 import {
   AgentTestContext,
   activityRepository,
@@ -29,15 +29,15 @@ describe('Agent Reply - /agents/:agentId/reply #novu-v2', () => {
       bridgeCalls.push(params);
     });
 
-    const chatSdkService = testServer.getService(ChatSdkService);
+    const outboundGateway = testServer.getService(OutboundGateway);
     sinon
-      .stub(chatSdkService, 'postToConversation')
+      .stub(outboundGateway, 'postToConversation')
       .resolves({ messageId: 'platform-msg-1', platformThreadId: 'platform-thread-1' });
     sinon
-      .stub(chatSdkService, 'editInConversation')
+      .stub(outboundGateway, 'editInConversation')
       .resolves({ messageId: 'platform-msg-1', platformThreadId: 'platform-thread-1' });
-    sinon.stub(chatSdkService, 'reactToMessage').resolves();
-    sinon.stub(chatSdkService, 'removeReaction').resolves();
+    sinon.stub(outboundGateway, 'reactToMessage').resolves();
+    sinon.stub(outboundGateway, 'removeReaction').resolves();
   });
 
   function postReply(body: Record<string, unknown>) {
@@ -288,7 +288,7 @@ describe('Agent Reply - /agents/:agentId/reply #novu-v2', () => {
   describe('addReactions', () => {
     it('should call reactToMessage for each addReaction entry', async () => {
       const conversationId = await seedConversation(ctx);
-      const chatSdkService = testServer.getService(ChatSdkService);
+      const outboundGateway = testServer.getService(OutboundGateway);
 
       const res = await postReply({
         conversationId,
@@ -300,13 +300,13 @@ describe('Agent Reply - /agents/:agentId/reply #novu-v2', () => {
       });
 
       expect(res.status).to.equal(200);
-      expect((chatSdkService.reactToMessage as sinon.SinonStub).callCount).to.equal(2);
+      expect((outboundGateway.reactToMessage as sinon.SinonStub).callCount).to.equal(2);
 
-      const firstCall = (chatSdkService.reactToMessage as sinon.SinonStub).getCall(0).args;
+      const firstCall = (outboundGateway.reactToMessage as sinon.SinonStub).getCall(0).args;
       expect(firstCall[4]).to.equal('msg-abc');
       expect(firstCall[5]).to.equal('thumbs_up');
 
-      const secondCall = (chatSdkService.reactToMessage as sinon.SinonStub).getCall(1).args;
+      const secondCall = (outboundGateway.reactToMessage as sinon.SinonStub).getCall(1).args;
       expect(secondCall[4]).to.equal('msg-def');
       expect(secondCall[5]).to.equal('check');
     });
