+++ apps/api/src/app/agents/e2e/agent-webhook.e2e.ts
@@ -9,10 +9,10 @@ import { testServer } from '@novu/testing';
 import { expect } from 'chai';
 import type { EmojiValue } from 'chat';
 import sinon from 'sinon';
+import { ChatInstanceRegistry } from '../conversation-runtime/ingress/chat-instance.registry';
 import { AgentConfigResolver } from '../services/agent-config-resolver.service';
 import { AgentInboundHandler, InboundReactionEvent } from '../services/agent-inbound-handler.service';
 import { AgentExecutionParams, BridgeExecutorService } from '../services/bridge-executor.service';
-import { ChatSdkService } from '../services/chat-sdk.service';
 import { AgentEventEnum } from '../shared/enums/agent-event.enum';
 import { AgentPlatformEnum } from '../shared/enums/agent-platform.enum';
 import {
@@ -59,9 +59,9 @@ async function pollFor<T>(
 }
 
 async function clearChatSdkInstances(): Promise<void> {
-  const chatSdkService = testServer.getService(ChatSdkService);
+  const registry = testServer.getService(ChatInstanceRegistry);
 
-  await chatSdkService.onModuleDestroy();
+  await registry.onModuleDestroy();
 }
 
 function mockEmoji(name: string): EmojiValue {
